feat: SCAN-968, refactor block detail module
